+++ client/ui/Post.global.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="post">
-    <div class="message">{{data.message}} <a class="action" @click="removePost">x</a></div>
+    <div class="message">{{data.message}} <a class="action" @click="removePost">x Delete</a></div>
   </div>
 </template>
 
@@ -11,7 +11,7 @@ export default {
   props: ['data'],
   methods: {
     removePost () {
-      console.log("meow");
+      console.log("waf");
       Meteor.call('posts.remove', this.data._id);
     }
   }
@@ -30,6 +30,8 @@ a {
     .message {
         // Mixin
         .padding(12px);
+
+        color: green;
     }
 }
 </style>