+++ packages/vue-component/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'akryum:vue-component',
-  version: '0.1.0',
+  version: '0.1.1',
   summary: 'VueJS single-file components that hot-reloads',
   git: 'https://github.com/Akryum/meteor-vue-component',
   documentation: 'README.md'
@@ -35,7 +35,7 @@ Package.onUse(function(api) {
   if(process.env.NODE_ENV === 'development') {
     api.use('ecmascript');
     api.use('akryum:vue@1.0.0');
-    api.mainModule('dev-client.js', 'client');
+    api.mainModule('client/dev-client.js', 'client');
   }
 });
 
@@ -47,6 +47,5 @@ Package.onTest(function(api) {
 });
 
 Npm.depends({
-  'vue-hot-reload-api': '1.3.2',
   'socket.io-client': '1.4.6'
 })